on/off script
I use a custom static profile rgb #000000 because the speaker turn on light after a power outage or reboot etc.

When my pc is off or suspended. I want it to be off.

+++ QR30_timeout-_simple_animation.sh
@@ -92,9 +92,16 @@ SEND_BRIGHTNESS[10]="2eaaec6b00070d0b02000064ff850000000000000000000000000000000
 # --- send hex to speaker ---
 send_hidraw() {
     local hex="$1"
+    
+    # skip if system is suspending
+    if [ -f /tmp/qr30-sleep.lock ]; then
+        echo "locked"
+        return
+    fi
     echo "$hex" | xxd -r -p | dd bs=64 count=1 conv=sync of=$(find_qr30_hidraw) >/dev/null 2>&1
-
 }
+
+
 
 #I don't query it because it may be needed to fire a stream of interrupts to receive it
 # remember to use the same number os steps otherwhise up and down or not, I don't care
